+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.10.0-rc.36",
+  sdkVersion: "3.10.0-rc.37",
   genVersion: "2.670.3",
-  userAgent: "speakeasy-sdk/typescript 3.10.0-rc.36 2.670.3 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.10.0-rc.37 2.670.3 2.0 @orq-ai/node",
 } as const;