+++ packages/orq-rc/RELEASES.md
@@ -2779,3 +2779,13 @@ Based on:
 - [typescript v3.10.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.36 - packages/orq-rc
+
+## 2025-08-01 09:45:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.595.1 (2.670.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.37] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.37 - packages/orq-rc